Testing: Update the YUI Test harness for Grunt
- Fails the build on test failure
- Splits out our custom Grunt tasks
- Update the relative require paths after moving the tasks from the
Gruntfile
